refactor(routes/endpoints): -> controller/services logic
