+++ api/server/routes/index.js
@@ -9,7 +9,7 @@ const tokenizer = require('./tokenizer');
 const auth = require('./auth');
 const keys = require('./keys');
 const oauth = require('./oauth');
-const { router: endpoints } = require('./endpoints');
+const endpoints = require('./endpoints');
 const plugins = require('./plugins');
 const user = require('./user');
 const config = require('./config');
